pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "physics-learning-aqt"
description = "Quantum circuit learning of differential equations using the AQT backend provider"
readme = "README.md"
version = "0.1.0"
requires-python = ">=3.10,<3.13"
license = { text = "Apace 2.0" }
keywords = ["quantum", "circuit learning", "aqt provider", "physics learning"]
authors = [
  { name = "Giorgio Tosti Balducci", email = "giorgio.tosti-balducci@pasqal.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

# always specify a version for each package
# to maintain consistency
dependencies = ["qiskit-aqt-provider==1.8.1", "pydantic>=2.0", "qiskit_algorithms"]

[tool.hatch.metadata]
allow-direct-references = true
allow-ambiguous-features = true

# add one or more extras in the dependencies
[project.optional-dependencies]
extras = []

[tool.hatch.envs.default]
features = ["extras"]
dependencies = [
  "pytest",
  "pytest-cov",
  "pytest-xdist",
  "pytest-httpx",
  "nbconvert",
  "ipykernel",
  "pre-commit",
  "black",
  "ruff",
]

[tool.hatch.envs.default.scripts]
test = "pytest -vvv --cov-report=term-missing --cov-config=pyproject.toml --cov=tests --ignore=./tests/test_examples.py {args}"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = """-vvv --cov-report=term-missing --cov-config=pyproject.toml  --cov=tests"""
xfail_strict = true
filterwarnings = [
  "ignore:Call to deprecated create function FieldDescriptor",
  "ignore:Call to deprecated create function Descriptor",
  "ignore:Call to deprecated create function EnumDescriptor",
  "ignore:Call to deprecated create function EnumValueDescriptor",
  "ignore:Call to deprecated create function FileDescriptor",
  "ignore:Call to deprecated create function OneofDescriptor",
  "ignore:distutils Version classes are deprecated.",
]

[tool.hatch.envs.docs]
dependencies = [
  "mkdocs",
  "mkdocs-material",
  "mkdocstrings",
  "mkdocstrings-python",
  "mkdocs-section-index",
  "mkdocs-exclude",
  "markdown-exec",
  "mike",
]

[tool.hatch.envs.docs.scripts]
build = "mkdocs build --clean --strict {args}"
serve = "mkdocs serve --dev-addr localhost:8000"

[[tool.hatch.envs.test.matrix]]
python = ["310", "311", "312"]

[tool.hatch.build.targets.wheel]
packages = ["physicslearningaqt"]

[tool.coverage.run]
branch = true
parallel = true
# uncomment to omit any file from the
# coverage. Regexps can be used
# to select all files from a folder
#omit = [
#  "template_python/to_omit.py",
#]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

[tool.ruff]
select = ["E", "F", "I", "Q"]
extend-ignore = ["F841"]
line-length = 100

[tool.ruff.isort]
required-imports = ["from __future__ import annotations"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.mccabe]
max-complexity = 15

[tool.ruff.flake8-quotes]
docstring-quotes = "double"

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = false
ignore_missing_imports = true
